+++ qiskit/providers/ibmq/credentials/__init__.py
@@ -40,16 +40,16 @@ def discover_credentials(
         3. in the `qiskitrc` configuration file
 
     Args:
-        qiskitrc_filename (str): location for the `qiskitrc` configuration
+        qiskitrc_filename: location for the `qiskitrc` configuration
             file. If `None`, defaults to `{HOME}/.qiskitrc/qiskitrc`.
 
     Returns:
-        dict: dictionary with the contents of the configuration file, with
+        dictionary with the contents of the configuration file, with
             the form::
 
             {credentials_unique_id: Credentials}
     """
-    credentials = OrderedDict()
+    credentials = OrderedDict()  # type: ignore[var-annotated]
 
     # dict[str:function] that defines the different locations for looking for
     # credentials, and their precedence order.
@@ -63,7 +63,7 @@ def discover_credentials(
     # Attempt to read the credentials from the different sources.
     for display_name, (reader_function, kwargs) in readers.items():
         try:
-            credentials = reader_function(**kwargs)
+            credentials = reader_function(**kwargs)  # type: ignore[arg-type]
             logger.info('Using credentials from %s', display_name)
             if credentials:
                 break