+++ test/ibmq/test_proxies.py
@@ -22,11 +22,12 @@ import sys
 
 from requests.exceptions import ProxyError
 
-from qiskit.providers.ibmq.api_v2.clients import (AuthClient,
-                                                  VersionClient)
-from qiskit.providers.ibmq.api_v2.exceptions import RequestsApiError
+from qiskit.providers.ibmq import IBMQFactory
+from qiskit.providers.ibmq.api.clients import (AuthClient,
+                                               VersionClient)
+from qiskit.providers.ibmq.api.exceptions import RequestsApiError
 from ..ibmqtestcase import IBMQTestCase
-from ..decorators import requires_qe_access, requires_new_api_auth
+from ..decorators import requires_qe_access
 
 # Fallback mechanism. Version variable is stored under __doc__ in new pproxy versions
 try:
@@ -66,9 +67,26 @@ class TestProxies(IBMQTestCase):
             self.proxy_process.wait()
 
     @requires_qe_access
-    @requires_new_api_auth
+    def test_proxies_factory(self, qe_token, qe_url):
+        """Should reach the proxy using factory.enable_account."""
+        factory = IBMQFactory()
+        provider = factory.enable_account(qe_token, qe_url,
+                                          proxies={'urls': VALID_PROXIES})
+
+        self.proxy_process.terminate()  # kill to be able of reading the output
+
+        auth_line = pproxy_desired_access_log_line(qe_url)
+        api_line = pproxy_desired_access_log_line(provider.credentials.url)
+        proxy_output = self.proxy_process.stdout.read().decode('utf-8')
+
+        # Check if the authentication call went through proxy.
+        self.assertIn(auth_line, proxy_output)
+        # Check if the API call (querying providers list) went through proxy.
+        self.assertIn(api_line, proxy_output)
+
+    @requires_qe_access
     def test_proxies_authclient(self, qe_token, qe_url):
-        """Should reach the proxy using IBMQClient."""
+        """Should reach the proxy using AuthClient."""
         pproxy_desired_access_log_line_ = pproxy_desired_access_log_line(qe_url)
 
         _ = AuthClient(qe_token, qe_url, proxies=VALID_PROXIES)
@@ -79,7 +97,6 @@ class TestProxies(IBMQTestCase):
 
     # pylint: disable=unused-argument
     @requires_qe_access
-    @requires_new_api_auth
     def test_proxies_versionclient(self, qe_token, qe_url):
         """Should reach the proxy using IBMQVersionFinder."""
         pproxy_desired_access_log_line_ = pproxy_desired_access_log_line(qe_url)
@@ -92,39 +109,32 @@ class TestProxies(IBMQTestCase):
                       self.proxy_process.stdout.read().decode('utf-8'))
 
     @requires_qe_access
-    @requires_new_api_auth
     def test_invalid_proxy_port_authclient(self, qe_token, qe_url):
         """Should raise RequestApiError with ProxyError using AuthClient."""
         with self.assertRaises(RequestsApiError) as context_manager:
             _ = AuthClient(qe_token, qe_url, proxies=INVALID_PORT_PROXIES)
 
-        self.assertIsInstance(context_manager.exception.original_exception,
-                              ProxyError)
+        self.assertIsInstance(context_manager.exception.__cause__, ProxyError)
 
     # pylint: disable=unused-argument
     @requires_qe_access
-    @requires_new_api_auth
     def test_invalid_proxy_port_versionclient(self, qe_token, qe_url):
         """Should raise RequestApiError with ProxyError using VersionClient."""
         with self.assertRaises(RequestsApiError) as context_manager:
             version_finder = VersionClient(qe_url, proxies=INVALID_PORT_PROXIES)
             version_finder.version()
 
-        self.assertIsInstance(context_manager.exception.original_exception,
-                              ProxyError)
+        self.assertIsInstance(context_manager.exception.__cause__, ProxyError)
 
     @requires_qe_access
-    @requires_new_api_auth
     def test_invalid_proxy_address_authclient(self, qe_token, qe_url):
         """Should raise RequestApiError with ProxyError using AuthClient."""
         with self.assertRaises(RequestsApiError) as context_manager:
             _ = AuthClient(qe_token, qe_url, proxies=INVALID_ADDRESS_PROXIES)
 
-        self.assertIsInstance(context_manager.exception.original_exception,
-                              ProxyError)
+        self.assertIsInstance(context_manager.exception.__cause__, ProxyError)
 
     @requires_qe_access
-    @requires_new_api_auth
     def test_invalid_proxy_address_versionclient(self, qe_token, qe_url):
         """Should raise RequestApiError with ProxyError using VersionClient."""
         # pylint: disable=unused-argument
@@ -133,8 +143,7 @@ class TestProxies(IBMQTestCase):
                                            proxies=INVALID_ADDRESS_PROXIES)
             version_finder.version()
 
-        self.assertIsInstance(context_manager.exception.original_exception,
-                              ProxyError)
+        self.assertIsInstance(context_manager.exception.__cause__, ProxyError)
 
 
 def pproxy_desired_access_log_line(url):